Code review changes for api
