+++ backend/models/postgis/project.py
@@ -133,6 +133,9 @@ class Project(db.Model):
     validation_permission = db.Column(
         db.Integer, default=ValidationPermission.ANY.value
     )  # Means only users with validator role can validate
+    enforce_assignment = db.Column(
+        db.Boolean, default=False
+    )  # Turns on task assignment
     enforce_random_task_selection = db.Column(
         db.Boolean, default=False
     )  # Force users to edit at random to avoid mapping "easy" tasks
@@ -358,6 +361,7 @@ class Project(db.Model):
         self.status = ProjectStatus[project_dto.project_status].value
         self.priority = ProjectPriority[project_dto.project_priority].value
         self.default_locale = project_dto.default_locale
+        self.enforce_assignment = project_dto.enforce_assignment
         self.enforce_random_task_selection = project_dto.enforce_random_task_selection
         self.private = project_dto.private
         self.mapper_level = MappingLevel[project_dto.mapper_level.upper()].value
@@ -968,6 +972,7 @@ class Project(db.Model):
         base_dto.validation_permission = ValidationPermission(
             self.validation_permission
         ).name
+        base_dto.enforce_assignment = self.enforce_assignment
         base_dto.enforce_random_task_selection = self.enforce_random_task_selection
         base_dto.private = self.private
         base_dto.mapper_level = MappingLevel(self.mapper_level).name
